Reorganized files, separated VolunteerHome and Calendar
