+++ src/pages/home/Home.css
@@ -1,0 +1,117 @@
+.calendar {
+  display: flex;
+  flex-direction: column;
+
+  flex: 1;
+
+  padding: 43px 27px 70px;
+  gap: 25px;
+}
+
+.calendarNav {
+  display: flex;
+  flex-direction: row;
+
+  justify-content: center;
+  align-items: center;
+
+  gap: 56px;
+
+  height: 43px;
+}
+
+.calendarNavButton {
+  border: 0px;
+  background-color: transparent;
+  color: var(--purple1);
+  font-size: 30px;
+  font-weight: 600;
+}
+
+.calendarNavIcon {
+  stroke: var(--purple1);
+  stroke-width: 4;
+  width: 30px;
+  height: 30px;
+}
+
+.timeContainer {
+  display: flex;
+  flex-direction: column;
+}
+
+.sunIcon {
+  stroke: var(--dark-gray);
+  width: 20px;
+  height: 20px;
+}
+
+.boldedSymbol {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  font-size: 18px;
+  font-weight: 800;
+  color: var(--dark-gray);
+  gap: 10px;
+}
+
+.timeList {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  flex: 1;
+  text-align: right;
+}
+
+.rightMainSection {
+  display: flex;
+  flex-direction: row;
+  gap: 25px;
+  flex: 1;
+}
+
+.gridContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 25px;
+  flex: 1;
+}
+
+.headerGrid {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  grid-template-rows: repeat(2, 28px);
+
+  align-items: center;
+}
+
+.headerGridItem {
+  text-align: center;
+
+  color: var(--dark-gray);
+}
+
+
+.calendarGrid {
+  display: grid;
+  grid-template-columns: repeat(7, 1fr);
+  grid-template-rows: repeat(20, 1fr);
+
+  border-radius: 10px;
+  outline: 1.5px solid var(--purple1);
+  outline-offset: -1px;
+
+  overflow: hidden;
+  flex: 1;
+}
+
+.calendarGridItemTop {
+  border: 1px solid var(--purple1);
+  border-bottom: 1px dashed var(--purple1);
+}
+
+.calendarGridItemBottom {
+  border: 1px solid var(--purple1);
+  border-top: 1px dashed var(--purple1);
+}
